requirements.txt
django
djangorestframework
mysqlclient  # Or psycopg2-binary if using PostgreSQL
python-decouple
drf_yasg